refactored security
